+++ Assets/_GameAssets/Scripts/Gameplay/Player/PlayerController.cs
@@ -29,6 +29,9 @@ public class PlayerController : MonoBehaviour
     [SerializeField] private float _slideSpeed;
     [SerializeField] private KeyCode _slideActivateKey;
 
+    [Header("Player")]
+    [SerializeField] private HealthManager _playerHealthManager;
+
     private Vector3 _movementDirection;
     private StateController _currentState;
     private bool isSliding = false;
@@ -116,11 +119,16 @@ public class PlayerController : MonoBehaviour
         else if (!isGrounded)
             newState = PlayerState.Jump;
 
+
         if (_currentState.GetPlayerState() != newState)
         {
             _currentState.SetPlayerState(newState);
             _state = newState;
             OnPlayerStateChanged?.Invoke(newState);
+            /*if (_currentState.GetPlayerState() == PlayerState.Jump) //HealthManagerControl
+                _playerHealthManager.Damage(1);
+            else if (_currentState.GetPlayerState() == PlayerState.Slide)
+                _playerHealthManager.InstantDead();    */
         }
     }